rewrote actionview using groups and spinners
