+++ .idea/workspace.xml
@@ -6,16 +6,13 @@
   <component name="ChangeListManager">
     <list default="true" id="dc124562-4597-4485-8f5f-be859a97acb8" name="Changes" comment="fixed actiongroup nonsense for file/about issue (clicking about button disabled buttons in file dropdown)">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/gui/toolbar.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/toolbar.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/file_toolbar_button.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/toolbar_buttons/file_toolbar_button.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/toolbar_dropdowns/file_dropdown.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/toolbar_buttons/toolbar_dropdowns/file_dropdown.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/toolbar_dropdowns/help_dropdown.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/toolbar_buttons/toolbar_dropdowns/help_dropdown.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/file_toolbar_button.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/help_toolbar_button.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/toolbar_dropdowns/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/toolbar_dropdowns/file_dropdown.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/toolbar_buttons/toolbar_dropdowns/help_dropdown.py" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/gui/toolbar_view.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/toolbar_view.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/logger/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/logger/__init__.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/logger/logger.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/logger/logger.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/main.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/translations/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/translations/__init__.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/translations/en_US.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/translations/en_US.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -30,6 +27,11 @@
     </option>
   </component>
   <component name="Git.Settings">
+    <option name="RECENT_BRANCH_BY_REPOSITORY">
+      <map>
+        <entry key="$PROJECT_DIR$" value="dropdown_popup_interactiob_bug" />
+      </map>
+    </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
   </component>
   <component name="GitHubPullRequestSearchHistory">{
@@ -63,7 +65,7 @@
     "Python.toolbar_view.executor": "Run",
     "RunOnceActivity.ShowReadmeOnStart": "true",
     "RunOnceActivity.git.unshallow": "true",
-    "git-widget-placeholder": "dynamic__button__dropdowns",
+    "git-widget-placeholder": "#1 on dynamic__button__dropdowns",
     "last_opened_file_path": "/home/aurora/PycharmProjects/pygenes",
     "settings.editor.selected.configurable": "preferences.lookFeel"
   }
