--- src/gui/toolbar_buttons/__init__.py
@@ -1,3 +1,0 @@
-from gui.toolbar_buttons.file_toolbar_button import FileToolbarButton
-from gui.toolbar_buttons.help_toolbar_button import HelpToolbarButton
-
