--- src/gui/toolbar_buttons/toolbar_dropdowns/__init__.py
@@ -1,2 +1,0 @@
-from gui.toolbar_buttons.toolbar_dropdowns.file_dropdown import FileDropdown
-from gui.toolbar_buttons.toolbar_dropdowns.help_dropdown import HelpDropdown
